Crate buny_test_runner

buny-test-runner

A tool for running tests on Buny sources via a custom source runner.

This features a (nearly) complete Buny source runner backed by wasmer, barring the following features:

send_partial_result: the functionality doesn’t make sense in tests, so we don’t need it.

net module rate limiting: I was lazy.

html mutating functions: unsure of how to implement this with the scraper library.

js module webview components: not sure how to implement this.

locale handling in parse_date: chrono doesn’t support this, and I’m not sure if there’s a good alternative.

However, I haven’t tested most of the functionality yet to be honest. Feel free to make an issue if you encounter any problems.

Usage

This crate depends on buny-test. Follow the instructions there to configure tests in your rust code.

You can also run the binary directly from the command line:

buny-test-runner <path_to_wasm_file>

Re-exports

pub use libs::FFIResult;

pub use libs::Ptr;

pub use libs::Rid;

pub use libs::WasmEnv;

Modules

imports

libs